+++ src/vector.c
@@ -49,13 +49,13 @@ void cutil_vector_clear(cutil_vector* vector) {
 	vector->_data = NULL;
 }
 
-unsigned int cutil_vector_size(cutil_vector* vector){
+size_t cutil_vector_size(cutil_vector* vector){
     return vector->_size;
 }
 
 void _grow_vector(cutil_vector* vector) {
 	if (vector->_size == vector->_capacity) {
-		unsigned int new_capacity = (vector->_capacity > 0) ? vector->_capacity * 2 : 1;
+		size_t new_capacity = (vector->_capacity > 0) ? vector->_capacity * 2U : 1U;
 
 		void* new_data = malloc(vector->_trait->size * new_capacity);
 		memcpy(new_data, vector->_data, vector->_size * vector->_trait->size);
@@ -72,7 +72,7 @@ void _grow_vector(cutil_vector* vector) {
 void cutil_vector_push(cutil_vector* vector, void* data) {
 	_grow_vector(vector);
 
-	unsigned int byte_offset = vector->_size * vector->_trait->size;
+	size_t byte_offset = vector->_size * vector->_trait->size;
 	vector->_trait->copy_func((char*)vector->_data + (byte_offset), data, vector->_trait->user_data);
 
 	vector->_size += 1;
@@ -88,7 +88,7 @@ void cutil_vector_pop(cutil_vector* vector) {
 
 bool cutil_vector_get(cutil_vector* vector, unsigned int index, void* out) {
 	if (index < vector->_size) {
-		unsigned int byte_offset = index * vector->_trait->size;
+		size_t byte_offset = index * vector->_trait->size;
 		memcpy(out, (char*)vector->_data + byte_offset, vector->_trait->size);
 
 		return true;
